It works!
Now correctly parses the player and goalie stats tables, fortunately handling any
weird data entered for player number (* or XX). It's likely still fragile, in that
it's doing a straight type conversion of things that may not be nums or integers, but
I think they seed the data with zeroes where appropriate, except for the player num,
so this may be safe.

+++ siahl_app/models.py
@@ -60,7 +60,7 @@ class PlayerStat(models.Model):
   # goalie stats
   shots   = models.IntegerField(null=True, blank=True, default=0)
   ga      = models.IntegerField(null=True, blank=True, default=0)
-  gaa     = models.IntegerField(null=True, blank=True, default=0)
+  gaa     = models.FloatField(null=True, blank=True, default=0.0)
   save_p  = models.FloatField(null=True, blank=True, default=0.0)
